Corrected a misleading declaration of the learning rate for weights
In earlier versions, learning rates for the weight matrix (lrate_wmat) were set to .01, but were actually not being applied -- in other words, the actual learning rate was 1.0. I've corrected the value when setting lrate_wmat to 1.0, and corrected it to actually be applied with learning, in case others want to play around with that parameter.